refactor(tests): move all custom g2p tests to one file
